pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remake-sdk"
version = "2.0.0a1"
description = "Remake.ai Robot SDK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Remake.ai"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.scripts]
remake = "remake_sdk.cli:main"

[project.optional-dependencies]
platform = [
    "python-socketio[client]>=5.0",
    "requests>=2.28",
]
socketio = [
    "python-socketio>=5.0",
]
podman = [
    "podman-py>=4.0",
]
mqtt = [
    "paho-mqtt>=2.0",
]
all = [
    "remake-sdk[platform,socketio,podman,mqtt]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/remake-ai/remake-sdk"
Documentation = "https://docs.remake.ai/sdk"

[tool.setuptools.packages.find]
where = ["."]
include = ["remake_sdk*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
